[refactor] 테스트 추가
