+++ src/test/java/step3/RacingCarTest.java
@@ -11,5 +11,9 @@ class RacingCarTest {
     @Test
     @DisplayName("움직였을 때 +1값을 출력한다.")
     void move() {
+        RacingCar car = new RacingCar();
+        car.move(9);
+        int preValue = car.getCurrentPosition();
+        assertThat(preValue).isEqualTo(1);
     }
 }
