+++ src/test/java/step3/RacingGameTest.java
@@ -1,15 +1,23 @@
-package study;
+package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
+import java.util.ArrayList;
+import java.util.List;
 
-public class RacingGameTest {
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+
+class RacingGame {
+
     @Test
-    @DisplayName("RacingGameTest 초기화 테스")트
+    @DisplayName("RacingGameTest 초기화 테스")
     void move() {
 
+        List<RacingCar> racingCarList = new ArrayList<>();
+        RacingGame racingGameTest = new RacingGame(2, racingCarList);
+
+        assertThat(racingCarList.size()).isEqualTo(2);
     }
 }
